Move compatibility code out of new client into mixin
